+++ fluidSimLib/volumeParticleEmitter3.h
@@ -17,10 +17,10 @@ public:
     volumeParticleEmitter3(
         const implicitSurface3Ptr& implicitSurface,
         const boundingBox3& bounds,
-        double spacing,
+        FloatType spacing,
         const vector3& initialVel = vector3(),
         size_t maxNumberOfParticles = std::numeric_limits<size_t>::max(),
-        double jitter = 0.0,
+        FloatType jitter = 0.0,
         bool isOneShot = true,
         bool allowOverlapping = false,
         uint32_t seed = 0);
@@ -28,9 +28,9 @@ public:
 
     void setPointGenerator(const pointGenerator3Ptr& newPointsGen);
 
-    double jitter() const;
+    FloatType jitter() const;
 
-    void setJitter(double newJitter);
+    void setJitter(FloatType newJitter);
 
     bool isOneShot() const;
 
@@ -44,10 +44,10 @@ public:
 
     void setMaxNumberOfParticles(size_t newMaxNumberOfParticles);
 
-    double spacing() const;
+    FloatType spacing() const;
 
     //! Sets the spacing between particles.
-    void setSpacing(double newSpacing);
+    void setSpacing(FloatType newSpacing);
 
     //! Sets the initial velocity of the particles.
     vector3 initialVelocity() const;
@@ -61,14 +61,14 @@ private:
 
     implicitSurface3Ptr mImplicitSurface;
     boundingBox3 mBounds;
-    double mSpacing;
+    FloatType mSpacing;
     vector3 mInitialVel;
     pointGenerator3Ptr mPointsGen;
 
     size_t mMaxNumberOfParticles = std::numeric_limits<size_t>::max();
     size_t mNumberOfEmittedParticles = 0;
 
-    double mJitter = 0.0;
+    FloatType mJitter = 0.0;
     bool mIsOneShot = true;
     bool mAllowOverlapping = false;
 
@@ -79,15 +79,15 @@ private:
     //! \param[in]  timeIntervalInSeconds   The time-step interval.
     //!
     void onUpdate(
-        double currentTimeInSeconds,
-        double timeIntervalInSeconds) override;
+        FloatType currentTimeInSeconds,
+        FloatType timeIntervalInSeconds) override;
 
     void emit(
         const particleSystemData3Ptr& particles,
         std::vector<vector3>* newPositions,
         std::vector<vector3>* newVelocities);
 
-    double random();
+    FloatType random();
 };
 
 typedef std::shared_ptr<volumeParticleEmitter3> volumeParticleEmitter3Ptr;
